exercicio 3

+++ strategy/ex3/src/Casa.java
@@ -1,6 +1,6 @@
 public class Casa extends Imovel{
 
-    Casa(int comodos, char localizacao, double espaco) {
+    Casa(int comodos, char localizacao, float espaco) {
         super(comodos, localizacao, espaco);
         setValorComportamento(new DefinirValorComEdificacao());
     }
